Set unmounted state For any left over timers.
Currently, we manually track it's mounted state. However, it doesn't actually track when it is unmounted.  This may lead to an actual issue. More importantly, it leads to log spam when the component is unmounted.

```
Warning: setState(...): Can only update a mounted or mounting component. This usually means you called setState() on an unmounted component. This is a no-op.
```

Unmounting is fairly popular for those of using `react-router` / etc, so it's nice to clean it up.

I don't exactly understand why we're not simply using the `isMounted` hook, but it was changed for 21134aecaa1cdbf06bb17ea9420a89ef81f0ca42 aka https://github.com/akiran/react-slick/pull/63/files#diff-3f43ed4258c814c46d337afc1a320ae2L252 .